Add files via upload
COOJA.testlog is the file where the mote output from the cooja simulator has to copied and pasted into it, and then run the try.c file to obtain the output of parent change count. NOTE: try.c and the testlog has to be in the same directory to execute seamlessly.